[Behavioral Analytics] Move overview components to folder
