+++ x-pack/plugins/enterprise_search/public/applications/analytics/components/analytics_collection_view/analytics_collection_overview/analytics_collection_chart.test.tsx
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { setMockValues } from '../../../__mocks__/kea_logic';
+import { setMockValues } from '../../../../__mocks__/kea_logic';
 
 import React from 'react';
 
@@ -16,7 +16,7 @@ import moment from 'moment';
 import { AreaSeries, Chart } from '@elastic/charts';
 import { EuiLoadingChart } from '@elastic/eui';
 
-import { FilterBy } from '../../utils/get_formula_by_filter';
+import { FilterBy } from '../../../utils/get_formula_by_filter';
 
 import { AnalyticsCollectionChart } from './analytics_collection_chart';
