+++ x-pack/plugins/enterprise_search/public/applications/analytics/components/analytics_collection_view/analytics_collection_overview/analytics_collection_metric.tsx
@@ -24,7 +24,7 @@ import { EuiThemeComputed } from '@elastic/eui/src/services/theme/types';
 import { i18n } from '@kbn/i18n';
 import { euiThemeVars } from '@kbn/ui-theme';
 
-import { withLensData } from '../../hoc/with_lens_data';
+import { withLensData } from '../../../hoc/with_lens_data';
 
 enum MetricStatus {
   INCREASE = 'increase',
